requirements_interview_detection.txt
# 面谈检测系统依赖包
# Interview Detection System Dependencies

# 核心深度学习框架
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0

# 数据处理和科学计算
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# 图像处理
Pillow>=8.3.0
opencv-python>=4.5.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 进度条和工具
tqdm>=4.62.0
wandb>=0.12.0  # 可选，用于实验跟踪

# 其他工具
dlib>=19.22.0  # 可选，用于面部特征点检测